+++ lib/plugins/rel/repository/JudgingMemberRepository.ts
@@ -1,4 +1,4 @@
-import { JudgingMember } from './JudgingMember'
+import { JudgingMember } from "../domain/JudgingMember";
 
 export interface JudgingMemberRepository {
     findById(userId: string): Promise<JudgingMember | undefined>
